Cell moved to standalone package.
