+++ aimmo-game/tests/test_simulation/test_world_map.py
@@ -6,7 +6,8 @@ from string import ascii_uppercase
 from unittest import TestCase
 
 from simulation.location import Location
-from simulation.world_map import Cell, WorldMap, world_map_static_spawn_decorator
+from simulation.world_map import WorldMap, world_map_static_spawn_decorator
+from simulation.maps.cell import Cell
 from .dummy_avatar import DummyAvatar
 from .maps import MockCell, MockPickup
